+++ app/models/folio/atom.rb
@@ -15,23 +15,9 @@ module Folio
       Base.recursive_subclasses
     end
 
-    def self.translations
-      @translations ||= begin
-        if Folio::Atom::Base.column_names.include?('title')
-          nil
-        else
-          Folio::Atom::Base.column_names
-                           .grep(/\Atitle_/)
-                           .map { |t| t.gsub(/\Atitle_/, '') }
-        end
-      end
-    end
-
     def self.text_fields
       @text_fields ||= begin
-        if Folio::Atom::Base.column_names.include?('title')
-          [:title, :content, :perex]
-        else
+        if ::Rails.application.config.folio_using_traco
           text_fields = []
           Folio::Atom::Base.column_names.each do |column|
             if column =~ /\A(title|content|perex)_/
@@ -39,6 +25,8 @@ module Folio
             end
           end
           text_fields
+        else
+          [:title, :content, :perex]
         end
       end
     end